+++ ali/neoforge/src/main/java/com/yanny/ali/neoforge/plugin/mods/farmers_delight/package-info.java
@@ -1,4 +1,0 @@
-@ParametersAreNonnullByDefault
-package com.yanny.ali.neoforge.plugin.mods.farmers_delight;
-
-import javax.annotation.ParametersAreNonnullByDefault;