+++ alicompat/common/src/main/java/com/yanny/alicompat/ICompatTranslations.java
@@ -1,0 +1,13 @@
+package com.yanny.alicompat;
+
+import org.jetbrains.annotations.NotNull;
+
+import java.util.Map;
+
+public interface ICompatTranslations {
+    @NotNull
+    String targetModId();
+
+    @NotNull
+    Map<String, String> getTranslations();
+}