+++ alicompat/fabric/src/main/java/com/yanny/alicompat/fabric/package-info.java
@@ -1,0 +1,4 @@
+@ParametersAreNonnullByDefault
+package com.yanny.alicompat.fabric;
+
+import javax.annotation.ParametersAreNonnullByDefault;