+++ alicompat/forge/src/main/java/com/yanny/alicompat/forge/platform/package-info.java
@@ -1,0 +1,4 @@
+@ParametersAreNonnullByDefault
+package com.yanny.alicompat.forge.platform;
+
+import javax.annotation.ParametersAreNonnullByDefault;