+++ alicompat/neoforge/src/compat/farmersdelight/java/com/yanny/alicompat/compat/farmersdelight/ReplaceItemModifierAccessor.java
@@ -1,0 +1,70 @@
+package com.yanny.alicompat.compat.farmersdelight;
+
+import com.yanny.aci.api.RangeValue;
+import com.yanny.aci.tooltip.TooltipBuilder;
+import com.yanny.ali.api.IDataNode;
+import com.yanny.ali.api.IItemNode;
+import com.yanny.ali.api.ILootModifier;
+import com.yanny.ali.api.IOperation;
+import com.yanny.ali.api.IServerUtils;
+import com.yanny.ali.plugin.common.NodeUtils;
+import com.yanny.ali.plugin.common.nodes.ItemNode;
+import com.yanny.ali.plugin.common.nodes.ModifiedNode;
+import com.yanny.ali.plugin.glm.GlobalLootModifierUtils;
+import com.yanny.ali.plugin.glm.IGlobalLootModifierAccessor;
+import com.yanny.ali.plugin.glm.ILootTableIdConditionPredicate;
+import com.yanny.ali.plugin.mods.BaseAccessor;
+import com.yanny.ali.plugin.mods.FieldAccessor;
+import com.yanny.ali.plugin.server.EnchantedRanges;
+import com.yanny.ali.plugin.server.TooltipUtils;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.level.storage.loot.entries.LootPoolSingletonContainer;
+import net.minecraft.world.level.storage.loot.predicates.LootItemCondition;
+import vectorwing.farmersdelight.common.loot.modifier.ReplaceItemModifier;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
+import java.util.Optional;
+import java.util.function.Function;
+import java.util.stream.Stream;
+
+public class ReplaceItemModifierAccessor extends BaseAccessor<ReplaceItemModifier> implements IGlobalLootModifierAccessor {
+    @FieldAccessor
+    private Item removedItem;
+    @FieldAccessor
+    private Item addedItem;
+    @FieldAccessor
+    private int addedCount;
+    @FieldAccessor
+    protected LootItemCondition[] conditions;
+
+    public ReplaceItemModifierAccessor(ReplaceItemModifier parent) {
+        super(parent);
+    }
+
+    public Optional<ILootModifier<?>> getLootModifier(IServerUtils utils, ILootTableIdConditionPredicate predicate) {
+        List<LootItemCondition> conditionList = Arrays.asList(this.conditions);
+
+        return GlobalLootModifierUtils.getLootModifier(conditionList, (c) -> {
+            Function<IDataNode, List<IDataNode>> factory = (src) -> {
+                List<IDataNode> nodes = new ArrayList<>();
+                IItemNode node = (IItemNode) src;
+                List<LootItemCondition> allConditions = Stream.concat(c.stream(), node.getConditions().stream()).toList();
+                EnchantedRanges chance = NodeUtils.getEnchantedChance(utils, allConditions, 1);
+                EnchantedRanges count = NodeUtils.getEnchantedCount(utils, Collections.emptyList());
+                TooltipBuilder tooltip = TooltipUtils.getTooltip(utils, LootPoolSingletonContainer.DEFAULT_QUALITY, chance, count, Collections.emptyList(), allConditions);
+
+                if (!c.isEmpty()) {
+                    nodes.add(new ModifiedNode(utils, src, new ItemNode(1, new RangeValue(addedCount), addedItem.getDefaultInstance(), tooltip.build(), Collections.emptyList(), allConditions)));
+                } else {
+                    nodes.add(new ItemNode(1, new RangeValue(addedCount), addedItem.getDefaultInstance(), tooltip.build(), Collections.emptyList(), allConditions));
+                }
+
+                return nodes;
+            };
+            return Collections.singletonList(new IOperation.ReplaceOperation((itemStack) -> itemStack.getItem().equals(removedItem), factory));
+        }, predicate);
+    }
+}