+++ alicompat/neoforge/src/main/java/com/yanny/alicompat/neoforge/datagen/package-info.java
@@ -1,0 +1,5 @@
+@ParametersAreNonnullByDefault
+package com.yanny.alicompat.neoforge.datagen;
+
+
+import javax.annotation.ParametersAreNonnullByDefault;