+++ awi/fabric/src/main/java/com/yanny/awi/fabric/platform/FabricPlatformHelper.java
@@ -1,5 +1,6 @@
 package com.yanny.awi.fabric.platform;
 
+import com.google.common.base.Suppliers;
 import com.yanny.aci.CommonLogUtils;
 import com.yanny.awi.Utils;
 import com.yanny.awi.api.IPlugin;
@@ -12,13 +13,21 @@ import org.slf4j.Logger;
 import java.nio.file.Path;
 import java.util.LinkedList;
 import java.util.List;
+import java.util.function.Supplier;
 
 public class FabricPlatformHelper implements IPlatformHelper {
     private static final Logger LOGGER = CommonLogUtils.getLogger(Utils.MOD_ID);
 
+    private final Supplier<List<IPlugin>> pluginsSupplier = Suppliers.memoize(this::loadPlugins);
+
     @NotNull
     @Override
     public List<IPlugin> getPlugins() {
+        return pluginsSupplier.get();
+    }
+
+    @NotNull
+    private List<IPlugin> loadPlugins() {
         List<IPlugin> plugins = new LinkedList<>();
 
         for (EntrypointContainer<IPlugin> container : FabricLoader.getInstance().getEntrypointContainers("awi", IPlugin.class)) {